pyproject.toml
[tool.poetry]
name = "saas"
version = "0.1.0"
description = ""
authors = ["Boris <boris@mail.com>"]

[tool.poetry.dependencies]
python = "^3.8.5"
uvicorn = "^0.13.4"
fastapi = "^0.63.0"
sqlalchemy = "^1.3.23"
alembic = "^1.5.2"
psycopg2-binary = "^2.8.6"
python-dotenv = "^0.17.0"
requests = "^2.25.1"
passlib = "^1.7.2"
argon2_cffi = "^20.1.0"
pg8000 = "^1.16.5"
python-jose = "^3.2.0"
python-multipart = "^0.0.5"
google-cloud-storage = "^1.37.1"
pillow = "^8.2.0"
email-validator = "^1.1.2"
websockets = "^8.1"
Faker = "^8.0.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
mypy = "^0.812"
flake8 = "^3.9.0"
black = "^20.8b1"
hypothesis = "^6.8.0"
behave = "^1.2.6"
selenium = "^3.141.0"
webdriver-manager = "^3.3.0"
ipython = "^7.21.0"

[tool.black]
line-lenghth = 119
skip-string-normalization = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--maxfail=1 -ra"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[mypy]
ignore_missing_imports = true
pretty = true
color_output = true
show_error_codes = true
error_summary = true
show_error_context = true
allow_untyped_globals = true
show_none_errors = false

[flake8]
max-line-length = 120
show-source = true
disable-noqa = false
benchmark = true
count = true
exclude = ["alembic"]
ignore = "F841, F811, W504, E123, E252, W503, E203, E501, E711"
